+++ .eslintrc.json
@@ -10,7 +10,6 @@
       "sourceType": "module"
   },
   "rules": {
-    "no-underscore-dangle": [ "error", { "allow": [ "_id" ] } ],
-    "max-classes-per-file": ["error", 5]
+    "no-underscore-dangle": [ "error", { "allow": [ "_id" ] } ]
   }
 }
