+++ README.md
@@ -1,6 +1,6 @@
 # news-explorer-api
 Доменные имена сервера: www.newsexplorerapi.students.nomoreparties.space и newsexplorerapi.students.nomoreparties.space
 
-Публичный IP сервера: 178.154.228.180
+Публичный IP сервера: 178.154.252.98
 
 Адрес сервера: https://www.newsexplorerapi.students.nomoreparties.space
