+++ middlewares/errors/ConflictError.js
@@ -1,0 +1,8 @@
+class ConflictError extends Error {
+  constructor(message) {
+    super(message);
+    this.statusCode = 409;
+  }
+}
+
+module.exports = { ConflictError };
