+++ middlewares/errors/NotFoundError.js
@@ -1,0 +1,8 @@
+class NotFoundError extends Error {
+  constructor(message) {
+    super(message);
+    this.statusCode = 404;
+  }
+}
+
+module.exports = { NotFoundError };
