+++ routes/article.js
@@ -2,7 +2,6 @@ const PostArticle = require('express').Router();
 const { celebrate, Joi } = require('celebrate');
 const { default: validator } = require('validator');
 const { getArticles, createArticle, deleteArticle } = require('../controllers/article');
-const { NotFoundError } = require('../middlewares/error');
 const auth = require('../middlewares/auth');
 
 PostArticle.use(auth);
@@ -34,9 +33,5 @@ PostArticle.delete('/articles/:articleId', celebrate({
     articleId: Joi.string().hex().length(24).required(),
   }),
 }), deleteArticle);
-
-PostArticle.use('/', () => {
-  throw new NotFoundError('Запрос не найден.');
-});
 
 module.exports = PostArticle;
