+++ routes/index.js
@@ -1,0 +1,8 @@
+const AllRoutes = require('express').Router();
+const PostUsers = require('./users');
+const PostArticle = require('./article');
+
+AllRoutes.use('/', PostUsers);
+AllRoutes.use('/', PostArticle);
+
+module.exports = AllRoutes;
